requirements.txt
asgiref==3.7.2
Django==5.0
django-extensions==3.2.3
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
